Use HW Decoders before 4.3 for a list of known devices
Add HWDecoderUtil: Utility class that return the preferred hardware decoder
(none, mediacodec and/or iomx) from a list of known devices. It's only taking
effect before Android 4.3 as MediaCodec will always be preferred after.

That class use the hidden android.os.SystemProperties API to retrieve
properties like "ro.board.platform" in order to know which device we are
running. That API won't be used after 4.3 as we want to limit the usage of
android private API.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>